+++ lite/tests/functional/detection/test_iou.py
@@ -1,24 +1,10 @@
 import numpy as np
-from valor_lite.detection import _compute_iou
+from valor_lite.detection import compute_iou
 
 
-def test__compute_iou():
+def test_compute_iou():
 
-    # datum id  0
-    # gt        1
-    # pd        2
-    # gt xmin   3
-    # gt xmax   4
-    # gt ymin   5
-    # gt ymax   6
-    # pd xmin   7
-    # pd xmax   8
-    # pd ymin   9
-    # pd ymax   10
-    # gt label  11
-    # pd label  12
-    # pd score  13
-
+    # xmin, xmax, ymin, ymax
     box1 = np.array([0.0, 10.0, 0.0, 10.0])
     box2 = np.array([5.0, 10.0, 0.0, 10.0])
     box3 = np.array([0.0, 5.0, 5.0, 10.0])
@@ -35,7 +21,7 @@ def test__compute_iou():
         ]
     )
 
-    ious = _compute_iou(pairs)
+    ious = compute_iou(pairs)
     assert len(ious) == 5
     assert ious[0] == 1.0
     assert ious[1] == 0.5
